Refactored LatelyPage and LatelyItem.
